autentificacion integrada

+++ routes/web.php
@@ -1,11 +1,13 @@
 <?php
 
+use App\Http\Controllers\AuthenticatedSessionController;
 use App\Http\Controllers\BannerController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\MiniaturaController;
 use App\Http\Controllers\NoticiaController;
 use App\Http\Controllers\ProgramacionController;
 use App\Http\Controllers\ProgramaController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -23,6 +25,7 @@ Route::get('/', function () {
     return view('welcome');
 });
 
+//rutas publicas
 Route::get('/', [HomeController::class, 'index'])->name('index');
 Route::get('/noticias/{slug}', [HomeController::class, 'detallenoticia'])->name('detallenoticia');
 Route::get('/quienessomos', [HomeController::class, 'quienessomos'])->name('quienessomos');
@@ -30,27 +33,35 @@ Route::get('/programacion', [HomeController::class, 'programacion'])->name('prog
 Route::get('/programas', [HomeController::class, 'programas'])->name('programas');
 Route::get('/contactos', [HomeController::class, 'contactos'])->name('contactos');
 
+Route::view('/login', 'auth.login')->name('login');
+Route::post('/login', [AuthenticatedSessionController::class, 'store']);
+Route::post('/logout', [AuthenticatedSessionController::class, 'destroy'])->name('logout');
+
 /* Rutas del administrador*/
 Route::get('/admin', function(){
     return view('admin.dashboard.index');
-});
+})->middleware('auth')->name('admin');
 
 Route::prefix('/admin')->group(function(){
-    Route::get('/programacion', [ProgramacionController::class, 'index'])->name('programacion.index');
-    Route::get('/programacion/create', [ProgramacionController::class, 'create'])->name('programacion.create');
-    Route::post('/programacion', [ProgramacionController::class, 'store'])->name('programacion.store');
-    Route::get('/programacion/{programacion}/edit', [ProgramacionController::class, 'edit'])->name('programacion.edit');
-    Route::patch('/programacion/{programacion}', [ProgramacionController::class, 'update'])->name('programacion.update');
-    Route::delete('/programacion/{programacion}', [ProgramacionController::class, 'destroy'])->name('programacion.destroy');
+    Route::get('/programacion', [ProgramacionController::class, 'index'])->middleware('auth')->name('programacion.index');
+    Route::get('/programacion/create', [ProgramacionController::class, 'create'])->middleware('auth')->name('programacion.create');
+    Route::post('/programacion', [ProgramacionController::class, 'store'])->middleware('auth')->name('programacion.store');
+    Route::get('/programacion/{programacion}/edit', [ProgramacionController::class, 'edit'])->middleware('auth')->name('programacion.edit');
+    Route::patch('/programacion/{programacion}', [ProgramacionController::class, 'update'])->middleware('auth')->name('programacion.update');
+    Route::delete('/programacion/{programacion}', [ProgramacionController::class, 'destroy'])->middleware('auth')->name('programacion.destroy');
 
     /*
     Route::get('/programas', [ProgramaController::class, 'index'])->name('programas.index');
     Route::get('/programas/create', [ProgramaController::class, 'create'])->name('programas.create');
     Route::post('/programas', [ProgramaController::class, 'store'])->name('programas.store');*/
-    Route::resource('programas', ProgramaController::class);
-    Route::resource('noticias', NoticiaController::class);
-    Route::resource('banners', BannerController::class);
-    Route::resource('miniaturas', MiniaturaController::class);
+    Route::resource('programas', ProgramaController::class)->middleware('auth');
+    Route::resource('noticias', NoticiaController::class)->middleware('auth');
+    Route::resource('banners', BannerController::class)->middleware('auth');
+    Route::resource('miniaturas', MiniaturaController::class)->middleware('auth');
 
-
+    Route::get('users', [UserController::class, 'index'])->middleware('auth')->name('users.index');
+    Route::get('users/create', [UserController::class, 'create'])->middleware('auth')->name('users.create');
+    Route::post('users', [UserController::class, 'store'])->middleware('auth')->name('users.store');
+    Route::delete('users/{user}', [UserController::class, 'destroy'])->middleware('auth')->name('users.destroy');
+    
 });
